{"schema":"decodex/commit/1","summary":"Extract live overlay focus rendering","authority":"manual"}
